+++ takumi/src/rendering/components/border.rs
@@ -348,5 +348,7 @@ impl BorderProperties {
         pixel
       },
     );
+
+    canvas.buffer_pool.release(mask);
   }
 }
